requirements.txt
django==4.2.6
psycopg2-binary==2.9.9
pytest-django==4.5.2
Pillow==10.1
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3
pytest~=7.4.2
drf-nested-routers==0.93.4